server/requirements.txt
boto3==1.42.49
celery==5.3.6
dj-database-url==3.1.1
django-filter==25.2
django-jazzmin==3.0.0
django-storages==1.14.6
djangorestframework==3.16.1
docker==7.0.0
kubernetes==35.0.0
python-dotenv==1.2.1
redis==5.0.1
whitenoise==6.11.0
psycopg2==2.9.11